move connection handling into its own file
